requirements.txt
streamlit==1.35.0
google-generativeai==0.5.2
python-dotenv==1.0.1
langchain==0.1.16
langchain-core==0.1.27
PyPDF2==3.0.1
chromadb==0.4.24
faiss-cpu==1.7.4
fsspec[http]==2024.2.0
scipy==1.13.1
urllib3==1.26.18
smart-open==6.4.0
numpy==1.26.4
